feat: decouple v1-sdk and v2-sdk deps
