+++ packages/commons/src/token/scale.ts
@@ -1,0 +1,11 @@
+import { ethers } from 'ethers';
+
+export const scale = (tokenDecimals: number) => {
+  const f = (value: number) => {
+    return ethers.BigNumber.from(
+      ethers.utils.parseUnits(value.toFixed(tokenDecimals), tokenDecimals),
+    );
+  };
+
+  return f;
+};
