+++ packages/v1-sdk/src/actions/token/getBalance.ts
@@ -1,30 +1,23 @@
 import { GetBalanceArgs } from '../types/actionArgTypes';
-import { getEthBalance } from './getEthBalance';
-import { getERC20Balance } from './getERC20Balance';
 import { AMMInfo } from '../../common/api/amm/types';
 import { getAmmInfo } from '../../common/api/amm/getAmmInfo';
+import { getEthBalance, getERC20Balance } from '@voltz-protocol/commons-v2';
 
 export const getBalance = async ({
   ammId,
   signer,
 }: GetBalanceArgs): Promise<number> => {
-  if (signer.provider === undefined) {
-    throw new Error('Signer provider not found');
-  }
-
   const walletAddress = await signer.getAddress();
-  const provider = signer.provider;
   const ammInfo: AMMInfo = await getAmmInfo(ammId);
 
   let currentBalance: number;
   if (ammInfo.isEth) {
-    currentBalance = await getEthBalance({ walletAddress, provider });
+    currentBalance = await getEthBalance(walletAddress, signer);
   } else {
     currentBalance = await getERC20Balance({
-      walletAddress,
-      provider,
       tokenAddress: ammInfo.underlyingTokenAddress,
-      tokenDecimals: ammInfo.underlyingTokenDecimals,
+      walletAddress,
+      subject: signer,
     });
   }
