+++ packages/v1-sdk/src/actions/token/index.ts
@@ -1,5 +1,3 @@
 export * from './approvePeriphery';
 export * from './getBalance';
 export * from './getAllowanceToPeriphery';
-export * from './getERC20Balance';
-export * from './getEthBalance';
