--- packages/v1-sdk/src/common/contract-generators/getERC20TokenContract.ts
@@ -1,29 +1,0 @@
-import { ethers } from 'ethers';
-
-export const getERC20TokenContract = (
-  tokenAddress: string,
-  providerOrSigner: ethers.providers.Provider | ethers.Signer,
-): ethers.Contract => {
-  const abi: string[] = [
-    `
-    function approve(address, uint256) external returns (bool)
-    `,
-    `
-    function balanceOf(address) external view returns (uint256)
-    `,
-    `
-    function allowance(address,address) external view returns (uint256)
-    `,
-    `
-    function approve(address,uint256) external returns (bool)
-    `,
-  ];
-
-  const contract: ethers.Contract = new ethers.Contract(
-    tokenAddress,
-    abi,
-    providerOrSigner,
-  );
-
-  return contract;
-};
