+++ packages/v1-sdk/src/common/contract-generators/index.ts
@@ -1,3 +1,2 @@
 export * from './getPeripheryContract';
 export * from './getMarginEngineContract';
-export * from './getERC20TokenContract';
