--- packages/v1-sdk/src/common/index.ts
@@ -1,2 +1,0 @@
-export * from './contract-generators/getERC20TokenContract';
-export * from './token/getERC20Allowance';
